Version 1.9 of the flyway database has been created (a table mapping company offer reviews has been added). Review entities have been created.
- An entity containing data about a given company offer review has been created.
- A repository with an adapter for basic management of the created entity was created.
- Exception handling was added when a user has an opinion on a given offer.
- DTO was added for reading and writing opinions.

+++ backend/src/main/java/com/m4zek/backend/repository/CompanyOfferRepository.java
@@ -8,7 +8,7 @@ import java.util.Optional;
 
 public interface CompanyOfferRepository {
 
-    Optional<CompanyOffer> findById(int id);
+    Optional<CompanyOffer> findById(long id);
 
     CompanyOffer save(CompanyOffer companyOffer);
